bug fixes with image upload

+++ Models/SpremeniProfilnoSliko.php
@@ -4,24 +4,25 @@ namespace Models;
 
 class SpremeniProfilnoSliko
 {
-    public function shraniProfilnoSliko(): array{
+    public function shraniProfilnoSliko(): array {
         $err = [];
         $param = $this->role();
         $dir = __DIR__ . '/../data/userImages/' . $_SESSION[$param]  . '/';
 
         $oldFile = glob($dir . 'profilna.*');
-        var_dump($oldFile);
-        if(!unlink($oldFile[0])){
-            $err[] = 'Napaka pri brisanju stare datoteke!';
-            return $err;
+        if (!empty($oldFile)) {
+            if (!unlink($oldFile[0])) {
+                $err[] = 'Napaka pri brisanju stare datoteke!';
+                return $err;
+            }
         }
 
-        if(!move_uploaded_file($_FILES['file']['tmp_name'], $dir . 'profilna.' . strtolower(pathinfo($_FILES['file']['name'], PATHINFO_EXTENSION)))){
-            $err[] = 'Napaka pri shranjevanju datoteke!';
-            return $err;
+        if (!move_uploaded_file($_FILES['file']['tmp_name'], $dir . 'profilna.' . strtolower(pathinfo($_FILES['file']['name'], PATHINFO_EXTENSION)))) {
+            return ['Napaka pri shranjevanju datoteke!'];
         }
         return $err;
     }
+
     private function role(): string{
         if(isset($_SESSION['id_ucitelja'])){
             return 'id_ucitelja';
